requirements.txt
acme==1.21.0
APScheduler==3.10.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==21.2.0
blinker==1.4
certbot==1.21.0
certbot-nginx==1.21.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==4.0.0
charset-normalizer==3.1.0
ConfigArgParse==1.5.3
configobj==5.0.6
cryptography==41.0.1
dbus-python==1.2.18
distro==1.7.0
#distro-info===1.1build1
Django==4.2.2
django-bootstrap-modal-forms==3.0.3
django-ckeditor==6.5.1
django-countries==7.5.1
django-encrypted-field==1.0.5
django-encrypted-model-fields==0.6.5
django-extensions==3.2.3
django-filter==23.2
django-htmx==1.14.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-multiselectfield==0.1.12
django-tables2==2.5.3
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-extra-fields==3.5.0
drf-test-generator==0.0.5
filetype==1.2.0
graphviz==0.20.1
gunicorn==20.1.0
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
josepy==1.10.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
msgpack==1.0.5
oauthlib==3.2.0
parsedatetime==2.6
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
pycryptodomex==3.18.0
pydot==1.4.2
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.7.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyrsistent==0.18.1
python-dotenv==0.19.2
pytz==2023.3
PyYAML==5.4.1
redis==4.5.5
requests==2.31.0
requests-toolbelt==0.9.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
texttable==1.6.4
typing_extensions==4.6.3
tzlocal==5.0.1
unattended-upgrades==0.1
urllib3==2.0.3
wadllib==1.3.6
websocket-client==1.2.3
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
